refactor: 기능별 함수 분리
